refactor auth context
