+++ src/hooks/auth.jsx
@@ -1,20 +1,11 @@
-import React, { useCallback, useContext, useEffect, useState } from 'react';
+import { useCallback, useContext } from 'react';
 import {
     getAuth,
     signOut as firebaseSignOut,
     signInWithEmailAndPassword,
 } from 'firebase/auth';
-import { HashLoader } from 'react-spinners';
 import { app } from '@scripts/firebase';
-
-const override = {
-    position: 'absolute',
-    top: '50%',
-    left: '50%',
-    transform: 'translateX(-50%) translateY(-50%)',
-};
-
-export const AuthContext = React.createContext();
+import { AuthContext } from '@contexts/AuthContext';
 
 export const useSession = () => {
     const { currentUser } = useContext(AuthContext);
@@ -56,37 +47,3 @@ export const useAuth = () => {
 
     return { signIn, signOut };
 };
-
-const AuthProvider = ({ children }) => {
-    const [currentUser, setCurrentUser] = useState();
-    const [pending, setPending] = useState(true);
-
-    useEffect(() => {
-        const unsub = getAuth().onAuthStateChanged((user) => {
-            if (user) {
-                setCurrentUser(user);
-            }
-            setPending(false);
-        });
-        return unsub;
-    }, []);
-
-    if (pending) {
-        return (
-            <HashLoader
-                cssOverride={override}
-                size={100}
-                color={'#123abc'}
-                loading={pending}
-            />
-        );
-    } else {
-        return (
-            <AuthContext.Provider value={{ currentUser }}>
-                {children}
-            </AuthContext.Provider>
-        );
-    }
-};
-
-export default AuthProvider;
